X_Shoe_analysis/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
praw==7.7.1
google-api-python-client==2.136.0
pytrends==4.9.2
python-dotenv==1.0.1
vaderSentiment==3.3.2
requests==2.32.3
python-pptx==0.6.23
pillow==10.4.0

